[cuda] Collapse the CUDA RMSNorm layer fork into a rms_norm whole-op
The previous shape of this change bound RMS normalization to the cuda
context through a per-backend Layer fork: a new directory
nntrainer/layers/cuda_layers/ whose only member, CudaRMSNormLayer,
duplicated the host layer's orchestration and carried the backend
routing inside its own forwarding -- exactly the *_gpu / cl_layers /
cuda_layers fork shape the architecture's collapse rule forbids.
Type-string state it produced: "rms_norm" resolved to the app-side
causallm layer on the cpu context and to CudaRMSNormLayer on the cuda
context, while the gpu context binds RMSNormLayerCl under the near-miss
string "rmsnorm" -- three classes for one operation. The fork's own
justification (the host fp16 sum-of-squares overflow: a wide residual
row squares past the fp16 max and zeroes the row) is a
shared-implementation argument, not a fork argument -- the
FP32-accumulation contract belongs in the op table, where every backend
states it once.

Mechanism -- the whole-op shape the element-wise decode ops already use:

* ComputeOps::rms_norm(in, out, gamma, epsilon, active_rows,
  row_offset): out = in * rsqrt(mean(in^2)+eps) * gamma over the row
  window; gamma is {1,1,1,width}; the FP32-accumulated sum of squares
  is named as the contract in the doc comment. NI-throw default.
* CpuComputeOps::rms_norm: the host composition the cpu context already
  runs -- the arch-dispatched rms_norm_wrt_width_fp32/fp16 intrinsics
  (the FP16 kernel accumulates its sum of squares in FP32 by its own
  documented design) plus the broadcast gamma multiply with the
  dtype-cast guard (gamma cloned to the activation dtype when stored at
  a different one), applied to the live window only.
* CudaComputeOps::rms_norm: the ENTIRE dispatch half of the former
  CudaRMSNormLayer moved verbatim -- the rows<=32 all-FP16
  device-kernel branch (cuda_rmsnorm_fp16) AND its own fused host
  fallback (the rmsnorm_rows template over the 4 dtype combos, after
  the finishIfAsync coherence drain). The kernel files
  cuda_rmsnorm.{h,cpp} are untouched.
* ONE neutral core layer nntrainer/layers/llm/rms_norm_layer.{h,cpp}
  (nntrainer::RMSNormLayer, type "rms_norm"): shape/step orchestration
  plus a single getOps()->rms_norm(...) call per batch step.
* Registration: the cuda context swaps CudaRMSNormLayer for the neutral
  layer (same (type, LAYER_RMSNORM) form, comment rewritten to the
  now-true dispatch statement); the cpu context registers the neutral
  layer next to the other promoted LLM layers.
* nntrainer/layers/cuda_layers/ is DELETED (layer .h/.cpp + meson)
  along with its nntrainer/meson.build element line. Net, this
  change-set SHRINKS: a 222-line layer fork becomes an op-table
  override.
* unittest_compute_ops_dispatch gains RmsNormDispatchesThroughAttached-
  ContextOps: attached-table dispatch counter, exact fp32 value check
  (constant rows, analytic rms, gamma scaling), and the
  (active_rows, row_offset) window honored -- rows outside it stay
  untouched.

Per-context equivalence:

* cuda: both halves of the former layer dispatch are preserved -- the
  device branch verbatim (same dtype/row gates, same pointers: the
  window views alias the arguments' own buffers at the row_offset 0
  every caller passes, same kernel) and the host fallback verbatim (the
  fused normalize*gamma FP32-accumulate loop) -- so cuda numerics are
  byte-identical. The override deliberately does NOT delegate its
  fallback to CpuComputeOps::rms_norm: the two host bodies compute the
  same math through different instruction sequences (fused scalar loop
  vs SIMD intrinsics + separate gamma multiply), and unifying them is
  follow-up material once measured.
* cpu: CpuComputeOps::rms_norm IS the former cpu layer's math -- the
  identical intrinsic calls and the identical guarded gamma multiply --
  so cpu numerics are preserved. With the neutral layer registered at
  context init, the app's later runtime registration of its own
  rms_norm becomes the documented re-registration no-op and the cpu
  context resolves to the neutral class.
* gpu: untouched. No "rms_norm" binding exists on the gpu context today
  and ClComputeOps has no rms_norm override yet, so registering the
  neutral layer there would only convert a clear create-time
  unregistered-type error into a runtime NI throw. The named follow-up
  owns the gpu swap: ClComputeOps::rms_norm forwarding to the existing
  rmsnorm_cl kernels, RMSNormLayerCl retirement, and the "rmsnorm"
  near-miss type string.

Disclosed decisions (each deliberate, none a silent behavior change):

* The Cpu impl is lifted from the cpu layer's incremental_forwarding
  (intrinsics + guarded multiply), NOT from the cuda fork's scalar
  fused-gamma fallback: the intrinsics satisfy the same
  FP32-accumulation rationale while preserving the cpu context's
  existing numerics AND its SIMD throughput -- the scalar loop would
  silently change cpu bytes and slow cpu prefill.
* gamma is requested at the model weight dtype (the cuda fork's
  policy), replacing the cpu layer's FP32 hardcode. For FP32-weight
  models the request is identical. For quantized bins that store norm
  weights as FP16, the FP32 hardcode positionally misreads the packed
  weight file from this tensor onward (and can overrun a tightly-sized
  mmap) -- the weight-dtype request is the load-layout correction. An
  FP32-gamma bin under an FP32 weight dtype still loads exactly as
  before, and the op's dtype-cast guard covers any gamma/activation
  mismatch.
* The gamma weight is requested trainable=true, matching the cpu layer
  (the cuda fork passed false); forward-inert either way -- training is
  rejected by calcDerivative.
* The skip_prefill gate is `from == 0 || (to - from) > 1` -- the merged
  SwiGLU/GeGLU semantics where a resumed multi-token prefill still
  counts as prefill. The cuda fork's `from == 0`-only variant differed
  on exactly that resumed edge; this aligns the norm with its sibling
  gates.
* Rows per batch step are channel*(to-from) (the cuda fork's formula);
  the cpu layer passed the step height only. Identical for every
  in-tree consumer (all channel==1); the formula is the one that stays
  correct if a channeled consumer appears.
* forwarding() delegates to incremental_forwarding over the full height
  (cuda-fork behavior); the cpu layer's forwarding was an empty body.
  Strictly more functional, unreachable for the incremental-only
  in-tree consumers.
* setProperty is lenient (parse epsilon/skip_prefill, silently drop the
  rest): load-time hints such as packed=false reach the layer and must
  not throw; the strict-throw variant would reject them.
* One token in the moved cuda region is not verbatim: the (unreachable
  in-tree) unsupported-dtype throw text now names
  CudaComputeOps::rms_norm instead of the class this commit deletes.

Three-axis note: structure ++ (the fork directory is gone, one class
per context under "rms_norm" on cpu+cuda, +1 dispatch unit test, the
change-set net shrinks); performance 0 by construction (cuda decode
keeps the device kernel behind the same gates, cpu keeps the SIMD
intrinsics, one virtual call per whole op per step); memory 0 (no new
allocation on any path; the dtype-cast gamma clone existed before on
the same path).

Verification: not built here (measurement host reserved); CI is the
oracle for this rung; post-cascade golden re-validation on the
assembled ladder pending.
