--- nntrainer/layers/cuda_layers/meson.build
@@ -1,7 +1,0 @@
-cuda_layer_sources = [
-  'cuda_rmsnorm_layer.cpp',
-]
-
-foreach s : cuda_layer_sources
-  nntrainer_sources += meson.current_source_dir() / s
-endforeach
