+++ nntrainer/layers/llm/meson.build
@@ -7,6 +7,7 @@ llm_layer_sources = [
   'qkv_layer.cpp',
   'geglu_layer.cpp',
   'swiglu_layer.cpp',
+  'rms_norm_layer.cpp',
   'lm_head.cpp', # untied host lm_head FC (no OpenCL)
   # tie_word_embedding's OpenCL lm_head GEMV is #if defined(ENABLE_OPENCL)-guarded
   # internally, so it compiles host-only without OpenCL (the FP32 unittest build).
@@ -23,6 +24,7 @@ llm_layer_headers = [
   'qkv_layer.h',
   'geglu_layer.h',
   'swiglu_layer.h',
+  'rms_norm_layer.h',
 ]
 
 # scalar_multiply's GPU variant uses OpenCL directly (cl_mem / ClContext), so it
