+++ nntrainer/tensor/cpu_backend/cpu_ops_table.h
@@ -391,6 +391,12 @@ public:
   // out = cap * act(in / cap) via host Tensor ops + ActiFunc (logit
   // soft-capping; act_type is an ActivationType cast to int).
   void softcap(const Tensor &in, Tensor &out, float cap, int act_type) override;
+  // out = in * rsqrt(mean(in^2)+eps) * gamma over the live rows, via the
+  // arch-dispatched width-wise rms_norm intrinsics (FP32-accumulated sum of
+  // squares in both the FP32 and FP16 kernels).
+  void rms_norm(const Tensor &in, Tensor &out, const Tensor &gamma,
+                float epsilon, unsigned int active_rows,
+                unsigned int row_offset) override;
 };
 
 } // namespace nntrainer
