+++ test/unittest/unittest_compute_ops_dispatch.cpp
@@ -41,6 +41,7 @@ struct CallCounters {
   std::atomic<int> scopy{0};
   std::atomic<int> scalar_mul{0};
   std::atomic<int> softcap{0};
+  std::atomic<int> rms_norm{0};
 };
 
 /**
@@ -97,6 +98,12 @@ public:
                int act_type) override {
     counters_->softcap++;
     real_->softcap(in, out, cap, act_type);
+  }
+  void rms_norm(const nntrainer::Tensor &in, nntrainer::Tensor &out,
+                const nntrainer::Tensor &gamma, float epsilon,
+                unsigned int active_rows, unsigned int row_offset) override {
+    counters_->rms_norm++;
+    real_->rms_norm(in, out, gamma, epsilon, active_rows, row_offset);
   }
 
 private:
@@ -380,6 +387,36 @@ TEST_F(ComputeOpsDispatchTest, SoftcapDispatchesThroughAttachedContextOps) {
   // algebraically-identical 2*sigmoid(2x)-1 form, so allow a few float ULPs.
   EXPECT_NEAR(out.getValue<float>(0, 0, 0, 0), 2.0f * std::tanh(0.5f), 1e-5f);
   EXPECT_NEAR(out.getValue<float>(0, 0, 0, 7), 2.0f * std::tanh(0.5f), 1e-5f);
+}
+
+/**
+ * @brief The RMS-normalization whole-op dispatches through the attached
+ *        ContextData ops, the Cpu impl computes the analytic value, and the
+ *        (active_rows, row_offset) window is honored — rows outside the
+ *        window stay untouched.
+ */
+TEST_F(ComputeOpsDispatchTest, RmsNormDispatchesThroughAttachedContextOps) {
+  nntrainer::Tensor in(1, 1, 3, 4);
+  nntrainer::Tensor out(1, 1, 3, 4);
+  nntrainer::Tensor gamma(1, 1, 1, 4);
+  in.setValue(2.0f);
+  gamma.setValue(2.0f);
+  out.setValue(0.0f);
+
+  in.setContextData(ct_data);
+  // Window: normalize only the first 2 of 3 rows.
+  in.getOps()->rms_norm(in, out, gamma, /*epsilon=*/0.0f, /*active_rows=*/2,
+                        /*row_offset=*/0);
+
+  EXPECT_GT(counters->rms_norm.load(), 0);
+  // Constant row: rms = sqrt(mean(2^2)) = 2, so each element normalizes to
+  // 1.0; gamma = 2 scales it to 2.0 (all powers of two — exact in fp32 on
+  // every arch path).
+  EXPECT_FLOAT_EQ(out.getValue<float>(0, 0, 0, 0), 2.0f);
+  EXPECT_FLOAT_EQ(out.getValue<float>(0, 0, 1, 3), 2.0f);
+  // Row 2 is outside the active window: still the 0.0 sentinel.
+  EXPECT_FLOAT_EQ(out.getValue<float>(0, 0, 2, 0), 0.0f);
+  EXPECT_FLOAT_EQ(out.getValue<float>(0, 0, 2, 3), 0.0f);
 }
 
 int main(int argc, char **argv) {
